scraper/requirements.txt
# Scraper Dependencies
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
lxml>=5.0.0
